Add stage mouseEnableStillEvents flag to disable emitting still mouse events
By setting this flag to `false`, the system will stop emitting fake `still` mouse events used to detect mouseOver mouseOut when the mouse is not moving but objects on the scene are moving or appearing / disappearing.

This is an optimization to allow the user to opt out of this feature if there is no need for it.